+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Additional metrics for consensus
 - Network confirmation before marking a dead peer as offline
+### Fixed
+- Accepting checkpoint blocks after (re)download
 
 ## [v2.7.1] - 2020-05-12
 ### Fixed